meta-agl-bsp: Fix rcar3 kernel module bbappends for linux-dummy
The usrmerge fixes added to the various rcar3 kernel bbappends turn
out to cause issues when using linux-dummy as the kernel as the
IC EG system containers do, add the required extra checking to avoid
issues.

Bug-AGL: SPEC-4703

Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28472
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account

+++ meta-agl-bsp/meta-rcar-gen3/recipes-kernel/kernel-module-gles/kernel-module-gles.bbappend
@@ -8,9 +8,11 @@ module_do_compile:prepend() {
 do_install:append() {
     # Work around upstream not using ${nonarch_base_libdir}/modules
     if ${@bb.utils.contains('DISTRO_FEATURES', 'usrmerge', 'true', 'false', d)}; then
-        install -d ${D}${nonarch_base_libdir}/
-        mv ${D}/lib/modules ${D}${nonarch_base_libdir}/
-        rm -rf ${D}/lib
+        if [ -d ${D}/lib/modules ]; then
+            install -d ${D}${nonarch_base_libdir}/
+            mv ${D}/lib/modules ${D}${nonarch_base_libdir}/
+            rm -rf ${D}/lib
+        fi
     fi
 }
