+++ meta-agl-bsp/meta-rcar-gen3/recipes-kernel/kernel-module-vspm/kernel-module-vspm.bbappend
@@ -1,8 +1,10 @@
 do_install:append() {
     # Work around upstream not using ${nonarch_base_libdir}/modules
     if ${@bb.utils.contains('DISTRO_FEATURES', 'usrmerge', 'true', 'false', d)}; then
-        install -d ${D}${nonarch_base_libdir}/
-        mv ${D}/lib/modules ${D}${nonarch_base_libdir}/
-        rm -rf ${D}/lib
+        if [ -d ${D}/lib/modules ]; then
+            install -d ${D}${nonarch_base_libdir}/
+            mv ${D}/lib/modules ${D}${nonarch_base_libdir}/
+            rm -rf ${D}/lib
+        fi
     fi
 }
